Enable server support for 0-RTT data
This previously existed, but only for QUIC.

There are some unfortunate shortcomings with the protocol design here:

Because the client must send 0-RTT data whether or not the server
accepts it or even the client hello, there must be several
disjoint methods for identifying and skipping these messages.  One
of these is in the record_layer.rs, and works by trial decryption.
Another happens if the server rejects the client's hello altogether,
and skips encrypted messages between the two client hellos.

The amount of data to skip is limited but -- because the design
appears to be defective -- the quantity is expressed (in
`max_early_data_size` provided with a ticket) in units of plaintext
bytes, but skipping data requires it in units of padded, tagged
ciphertext bytes.  The server cannot compute one from the other,
so we're interpretting `max_early_data_size` as both at the same time.